+++ graphs.h
@@ -32,7 +32,7 @@
 // or not in the configuration. This bit may interfere with libraries using 8-bit 
 // alpha channels.
 //
-#define DEFCOLOR ((u_long) 0x80000000ul)
+#define DEFCOLOR ((uint64_t) 0x80000000ul)
 
 class config_t;
 class history_t;
@@ -49,22 +49,22 @@ class graph_t {
       int font_size_medium_bold_px;
 
       /* colors */
-      u_int black, white, dkgrey, red, blue, orange, green, cyan, yellow, purple, ltpurple, ltgreen, brown;
-      u_int c_shadow, c_background, c_gridline, c_hits, c_files, c_hosts, c_pages, c_visits, c_volume, c_outline, c_legend, c_weekend;
+      uint32_t black, white, dkgrey, red, blue, orange, green, cyan, yellow, purple, ltpurple, ltgreen, brown;
+      uint32_t c_shadow, c_background, c_gridline, c_hits, c_files, c_hosts, c_pages, c_visits, c_volume, c_outline, c_legend, c_weekend;
 
-      u_int graph_background;
-      u_int graph_gridline;
-      u_int graph_shadow;
-      u_int graph_title_color;
-      u_int graph_hits_color;
-      u_int graph_files_color;
-      u_int graph_hosts_color;
-      u_int graph_pages_color;
-      u_int graph_visits_color;
-      u_int graph_volume_color;
-      u_int graph_outline_color;
-      u_int graph_legend_color;
-      u_int graph_weekend_color;
+      uint32_t graph_background;
+      uint32_t graph_gridline;
+      uint32_t graph_shadow;
+      uint32_t graph_title_color;
+      uint32_t graph_hits_color;
+      uint32_t graph_files_color;
+      uint32_t graph_hosts_color;
+      uint32_t graph_pages_color;
+      uint32_t graph_visits_color;
+      uint32_t graph_volume_color;
+      uint32_t graph_outline_color;
+      uint32_t graph_legend_color;
+      uint32_t graph_weekend_color;
 
       gdImagePtr im;                      /* image buffer        */
       FILE *out;                          /* output file for PNG */
@@ -84,29 +84,29 @@ class graph_t {
       void draw_grid_line(gdImagePtr im, u_int gw, u_int y);
       void draw_graph_bar(gdImagePtr im, u_int x1, u_int y1, u_int x2, u_int y2, int color);
 
-      static bool is_default_color(u_long color);
+      static bool is_default_color(uint64_t color);
 
    public:
       graph_t(const config_t& _config);
 
       ~graph_t(void);
 
-      static u_long make_color(const char *str);
+      static uint32_t make_color(const char *str);
 
       // these methods should be called only after the engine has been initialized
-      u_long get_background_color(void) const {return graph_background;}
-      u_long get_title_color(void) const {return graph_title_color;}
-      u_long get_gridline_color(void) const {return graph_gridline;}
-      u_long get_shadow_color(void) const {return graph_shadow;}
-      u_long get_hits_color(void) const {return graph_hits_color;}
-      u_long get_files_color(void) const {return graph_files_color;}
-      u_long get_pages_color(void) const {return graph_pages_color;}
-      u_long get_visits_color(void) const {return graph_visits_color;}
-      u_long get_hosts_color(void) const {return graph_hosts_color;}
-      u_long get_volume_color(void) const {return graph_volume_color;}
-      u_long get_outline_color(void) const {return graph_outline_color;}
-      u_long get_legend_color(void) const {return graph_legend_color;}
-      u_long get_weekend_color(void) const {return graph_weekend_color;}
+      uint64_t get_background_color(void) const {return graph_background;}
+      uint64_t get_title_color(void) const {return graph_title_color;}
+      uint64_t get_gridline_color(void) const {return graph_gridline;}
+      uint64_t get_shadow_color(void) const {return graph_shadow;}
+      uint64_t get_hits_color(void) const {return graph_hits_color;}
+      uint64_t get_files_color(void) const {return graph_files_color;}
+      uint64_t get_pages_color(void) const {return graph_pages_color;}
+      uint64_t get_visits_color(void) const {return graph_visits_color;}
+      uint64_t get_hosts_color(void) const {return graph_hosts_color;}
+      uint64_t get_volume_color(void) const {return graph_volume_color;}
+      uint64_t get_outline_color(void) const {return graph_outline_color;}
+      uint64_t get_legend_color(void) const {return graph_legend_color;}
+      uint64_t get_weekend_color(void) const {return graph_weekend_color;}
 
       // return true if the respective color is defined in the configuration
       bool is_default_background_color(void) const {return config.graph_background & DEFCOLOR ? true : false;}
@@ -127,7 +127,7 @@ class graph_t {
       void cleanup_graph_engine(void);
 
       int year_graph6x(const history_t& history, const char *fname, const char *title, u_int& graph_width, u_int& graph_height);
-      int pie_chart(const char *fname, const char *title, u_long t_val, const u_long data1[], const char *legend[]);
+      int pie_chart(const char *fname, const char *title, uint64_t t_val, const uint64_t data1[], const char *legend[]);
       int month_graph6(const char *fname, const char *title, int month, int year, const daily_t daily[31]);
       int day_graph3(const char *fname, const char *title, const hourly_t hourly[24]);
 };
